apps/SMS_Scam_Detection/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
huggingface_hub